requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
pillow>=9.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Tracking
filterpy>=1.4.5
lap>=0.4.0

# OCR
easyocr>=1.7.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0
wandb>=0.15.0

# Evaluation
scikit-learn>=1.3.0
motmetrics>=1.4.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
jupyter>=1.0.0

# Hugging Face
huggingface_hub>=0.19.0